Move MessageClient to its own library
