+++ Source/CMakeLists.txt
@@ -26,6 +26,8 @@ option(WEBSOCKET
         "Include the websocket library." ON)
 option(TRACING
         "Include the tracing library." ON)
+option(MESSAGING
+        "Include the messaging library." ON)
 option(PROFILER
         "Include the profiler library." OFF)
 option(COM
@@ -111,6 +113,10 @@ endif()
 
 if(TRACING)
     add_subdirectory(tracing)
+endif()
+
+if(MESSAGING)
+    add_subdirectory(messaging)
 endif()
 
 if(WARNING_REPORTING)
