+++ Source/core/MessageUnit.cpp
@@ -101,7 +101,7 @@ namespace Core {
     }
     uint32_t MessageUnit::Open(const string& pathName)
     {
-        string basePath = Core::Format("%s/%s", pathName.c_str(), _T("MessageDispatcher"));
+        string basePath = Core::Format("%s%s", pathName.c_str(), _T("MessageDispatcher"));
         string identifier = _T("md");
 
         if (Core::File(basePath).IsDirectory()) {
@@ -212,118 +212,6 @@ namespace Core {
         Core::SafeSyncType<Core::CriticalSection> guard(_adminLock);
         auto entry = _controls.find({ type, category });
         _controls.erase(entry);
-    }
-
-    //====================================================================
-    MessageClient::MessageClient(const string& identifer, const string& basePath)
-        : _identifier(identifer)
-        , _basePath(basePath)
-    {
-    }
-
-    void MessageClient::AddInstance(uint32_t id)
-    {
-        _adminLock.Lock();
-
-        _clients.emplace(std::piecewise_construct,
-            std::forward_as_tuple(id),
-            std::forward_as_tuple(_identifier, id, false, _basePath));
-
-        _listId.emplace_back(id);
-
-        _adminLock.Unlock();
-    }
-
-    void MessageClient::RemoveInstance(uint32_t id)
-    {
-        Core::SafeSyncType<Core::CriticalSection> guard(_adminLock);
-        _clients.erase(id);
-    }
-
-    const std::list<uint32_t>& MessageClient::InstanceIds() const
-    {
-        Core::SafeSyncType<Core::CriticalSection> guard(_adminLock);
-        return _listId;
-    }
-
-    void MessageClient::WaitForUpdates(const uint32_t waitTime)
-    {
-        _adminLock.Lock();
-
-        if (!_clients.empty()) {
-            //ring is same for all dispatchers
-            auto firstEntry = _clients.begin();
-            _adminLock.Unlock();
-
-            firstEntry->second.Wait(waitTime);
-        } else {
-            _adminLock.Unlock();
-        }
-    }
-
-    void MessageClient::CancelWaiting()
-    {
-        _adminLock.Lock();
-
-        if (!_clients.empty()) {
-            //ring is same for all dispatchers
-            auto firstEntry = _clients.begin();
-            _adminLock.Unlock();
-
-            firstEntry->second.Ring();
-        } else {
-            _adminLock.Unlock();
-        }
-    }
-
-    void MessageClient::Enable(const bool enable, Core::MessageInformation::MessageType type, const string& category, const string& module = "MODULE_UNKNOWN")
-    {
-    }
-    bool MessageClient::IsEnabled(Core::MessageInformation::MessageType type, const string& category, const string& module = "MODULE_UNKNOWN") const
-    {
-        return false;
-    }
-
-    std::pair<MessageInformation, Core::ProxyType<IMessageEvent>> MessageClient::Pop(uint32_t id)
-    {
-        _adminLock.Lock();
-
-        uint8_t buffer[DataSize];
-        uint16_t size = sizeof(buffer);
-
-        MessageInformation information;
-        Core::ProxyType<IMessageEvent> message;
-
-        auto client = _clients.find(id);
-
-        if (client != _clients.end()) {
-
-            if (client->second.PopData(size, buffer) != Core::ERROR_NONE) {
-                //warning trace here
-            } else {
-                auto length = information.Deserialize(buffer, size);
-
-                auto factory = _factories.find(information.Type());
-                if (factory != _factories.end()) {
-                    message = factory->second->Create();
-                    message->Deserialize(buffer + length, size - length);
-                }
-            }
-        }
-        _adminLock.Unlock();
-
-        return { information, message };
-    }
-
-    void MessageClient::AddFactory(MessageInformation::MessageType type, Core::IMessageEventFactory* factory)
-    {
-        Core::SafeSyncType<Core::CriticalSection> guard(_adminLock);
-        _factories.emplace(type, factory);
-    }
-    void MessageClient::RemoveFactory(MessageInformation::MessageType type)
-    {
-        Core::SafeSyncType<Core::CriticalSection> guard(_adminLock);
-        _factories.erase(type);
     }
 
 }
